Move duplicate functions to LocalOscillator
