+++ src/qibolab/instruments/erasynth.py
@@ -11,11 +11,10 @@ import requests
 from qcodes_contrib_drivers.drivers.ERAInstruments import ERASynthPlusPlus
 from qibo.config import log
 
-from qibolab.instruments.abstract import InstrumentException
-from qibolab.instruments.oscillator import LocalOscillator, LocalOscillatorSettings
+from qibolab.instruments.oscillator import DummyDevice, LocalOscillator
 
 
-class ERASynthEthernet:
+class ERASynthEthernet(DummyDevice):
     MAX_RECONNECTION_ATTEMPTS = 10
     TIMEOUT = 10
 
@@ -25,6 +24,10 @@ class ERASynthEthernet:
         self.post("readAll", 1)
         self.post("readDiagnostic", 0)
         self.post("rfoutput", 0)
+
+    @property
+    def url(self):
+        return f"http://{self.address}/"
 
     def post(self, name, value):
         """Post a value to the instrument's web server.
@@ -38,7 +41,7 @@ class ERASynthEthernet:
         value = str(value)
         for _ in range(self.MAX_RECONNECTION_ATTEMPTS):
             try:
-                response = requests.post(f"http://{self.address}/", data={name: value}, timeout=self.TIMEOUT)
+                response = requests.post(self.url, data={name: value}, timeout=self.TIMEOUT)
                 if response.status_code == 200:
                     return True
                 break
@@ -56,7 +59,7 @@ class ERASynthEthernet:
         """
         for _ in range(self.MAX_RECONNECTION_ATTEMPTS):
             try:
-                response = requests.post(f"http://{self.address}/", params={"readAll": 1}, timeout=self.TIMEOUT)
+                response = requests.post(self.url, params={"readAll": 1}, timeout=self.TIMEOUT)
                 if response.status_code == 200:
                     # reponse.text is a dictonary in string format, convert it to a dictonary
                     return json.loads(response.text)[name]
@@ -88,83 +91,34 @@ class ERASynthEthernet:
 
 class ERA(LocalOscillator):
     def __init__(self, name, address, ethernet=True, reference_clock_source="int"):
-        super().__init__(name, address)
-        self.device = None
-        self.settings = LocalOscillatorSettings()
+        super().__init__(name, address, reference_clock_source)
         self.ethernet = ethernet
-        self._reference_clock_source = reference_clock_source
 
-    @property
-    def frequency(self):
-        return self.settings.frequency
+    def create(self):
+        if self.ethernet:
+            return ERASynthEthernet(self.name, self.address)
+        else:
+            return ERASynthPlusPlus(f"{self.name}", f"TCPIP::{self.address}::INSTR")
 
-    @frequency.setter
+    @LocalOscillator.frequency.setter
     def frequency(self, x):
         if self.frequency != x:
             self.settings.frequency = x
             if self.is_connected:
                 self.device.set("frequency", int(x))
 
-    @property
-    def power(self):
-        return self.settings.power
-
-    @power.setter
+    @LocalOscillator.power.setter
     def power(self, x):
         if self.power != x:
             self.settings.power = x
             if self.is_connected:
                 self.device.set("power", float(x))
 
-    @property
-    def reference_clock_source(self):
-        return self._reference_clock_source
-
-    @reference_clock_source.setter
+    @LocalOscillator.reference_clock_source.setter
     def reference_clock_source(self, x):
-        if self.reference_clock_source != x:
-            self._reference_clock_source = x
-            if self.is_connected:
-                self.device.ref_osc_source(x)
-
-    def upload(self):
-        if self.settings.frequency is not None:
-            self.device.set("frequency", self.settings.frequency)
-        if self.settings.power is not None:
-            self.device.set("power", self.settings.power)
-
-    def connect(self):
-        """Connects to the instrument using the IP address set in the runcard."""
-        if not self.is_connected:
-            for attempt in range(3):
-                try:
-                    if self.ethernet:
-                        self.device = ERASynthPlusPlus(self.name, self.address)
-                    else:
-                        self.device = ERASynthPlusPlus(f"{self.name}", f"TCPIP::{self.address}::INSTR")
-                    self.is_connected = True
-                    break
-                except KeyError as exc:
-                    log.info(f"Unable to connect:\n{str(exc)}\nRetrying...")
-                    self.name += "_" + str(attempt)
-                except ConnectionError as exc:
-                    log.info(f"Unable to connect:\n{str(exc)}\nRetrying...")
-            if not self.is_connected:
-                raise InstrumentException(self, f"Unable to connect to {self.name}")
-        else:
-            raise RuntimeError("There is an open connection to the instrument already")
-        self.upload()
-
-    def start(self):
-        self.device.on()
-
-    def stop(self):
-        self.device.off()
-
-    def disconnect(self):
+        self._reference_clock_source = x
         if self.is_connected:
-            self.device.close()
-            self.is_connected = False
+            self.device.ref_osc_source(x)
 
     def __del__(self):
         self.disconnect()
