+++ src/qibolab/instruments/oscillator.py
@@ -1,7 +1,13 @@
 from dataclasses import dataclass
 from typing import Optional
 
-from qibolab.instruments.abstract import Instrument, InstrumentSettings
+from qibo.config import log
+
+from qibolab.instruments.abstract import (
+    Instrument,
+    InstrumentException,
+    InstrumentSettings,
+)
 
 
 @dataclass
@@ -11,6 +17,8 @@ class LocalOscillatorSettings(InstrumentSettings):
 
 
 class DummyDevice:
+    ref_osc_source = None
+
     def set(self, name, value):
         """Set device property."""
 
@@ -33,10 +41,11 @@ class LocalOscillator(Instrument):
     They cannot be used to play or sweep pulses.
     """
 
-    def __init__(self, name, address):
+    def __init__(self, name, address, reference_clock_source="EXT"):
         super().__init__(name, address)
         self.device = None
         self.settings = LocalOscillatorSettings()
+        self._reference_clock_source = reference_clock_source
 
     @property
     def frequency(self):
@@ -44,7 +53,10 @@ class LocalOscillator(Instrument):
 
     @frequency.setter
     def frequency(self, x):
-        self.settings.frequency = x
+        if self.frequency != x:
+            self.settings.frequency = x
+            if self.is_connected:
+                self.device.set("frequency", x)
 
     @property
     def power(self):
@@ -52,15 +64,54 @@ class LocalOscillator(Instrument):
 
     @power.setter
     def power(self, x):
-        self.settings.power = x
+        if self.power != x:
+            self.settings.power = x
+            if self.is_connected:
+                self.device.set("power", x)
+
+    @property
+    def reference_clock_source(self):
+        return self._reference_clock_source
+
+    @reference_clock_source.setter
+    def reference_clock_source(self, x):
+        self._reference_clock_source = x
+        if self.is_connected:
+            self.device.ref_osc_source = x
+
+    def create(self):
+        """Create instance of physical device."""
+        return DummyDevice()
+
+    def connect(self):
+        """Connects to the instrument using the IP address set in the runcard."""
+        if not self.is_connected:
+            for attempt in range(3):
+                try:
+                    self.device = self.create()
+                    self.is_connected = True
+                    break
+                except KeyError as exc:
+                    log.info(f"Unable to connect:\n{str(exc)}\nRetrying...")
+                    self.name += "_" + str(attempt)
+                except ConnectionError as exc:
+                    log.info(f"Unable to connect:\n{str(exc)}\nRetrying...")
+            if not self.is_connected:
+                raise InstrumentException(self, f"Unable to connect to {self.name}")
+        else:
+            raise InstrumentException(self, "There is an open connection to the instrument already")
+        self.upload()
 
     def upload(self):
         """Uploads cached setting values to the instruments."""
+        if not self.is_connected:
+            raise InstrumentException(self, "Cannot upload settings if instrument is not connected.")
 
-    def connect(self):
-        self.is_connected = True
-        self.device = DummyDevice()
-        self.upload()
+        if self.settings.frequency is not None:
+            self.device.set("frequency", self.settings.frequency)
+        if self.settings.power is not None:
+            self.device.set("power", self.settings.power)
+        self.ref_osc_source = self._reference_clock_source
 
     def setup(self, **kwargs):
         """Update instrument settings.
@@ -85,4 +136,6 @@ class LocalOscillator(Instrument):
         self.device.off()
 
     def disconnect(self):
-        self.is_connected = False
+        if self.is_connected:
+            self.device.close()
+            self.is_connected = False
